requirements.txt
aioredis==1.3.0
amqp==1.4.9
anyjson==0.3.3
asgiref==3.1.0
asn1crypto==1.0.1
async-timeout==2.0.1
asyncio==3.4.3
attrs==19.2.0
autobahn==19.10.1
Automat==0.7.0
backcall==0.1.0
bcrypt==3.1.7
billiard==3.3.0.23
celery==3.1.25
cffi==1.12.3
channels==2.1.5
channels-redis==2.3.1
constantly==15.1.0
cryptography==2.7
daphne==2.3.0
decorator==4.4.0
Django==1.11.3
django-celery==3.2.2
django-cors-headers==2.1.0
django-filter==2.1.0
django-ranged-response==0.2.0
django-simple-captcha==0.5.11
djangorestframework==3.9.4
djangorestframework-jwt==1.11.0
gunicorn==19.9.0
hiredis==1.0.0
hyperlink==19.0.0
idna==2.8
importlib-metadata==1.1.0
include==0.2.2
incremental==17.5.0
ipdb==0.12
ipython==7.5.0
ipython-genutils==0.2.0
jedi==0.14.0
kombu==3.0.37
Markdown==3.1.1
more-itertools==8.0.0
msgpack==0.5.6
mysqlclient==1.3.10
numpy==1.16.4
opencv-python==4.1.0.25
pandas==0.25.3
paramiko==2.6.0
parso==0.5.0
pexpect==4.7.0
pickleshare==0.7.5
Pillow==6.0.0
prompt-toolkit==2.0.9
ptyprocess==0.6.0
pycparser==2.19
Pygments==2.4.2
PyHamcrest==1.9.0
PyJWT==1.7.1
PyNaCl==1.3.0
python-dateutil==2.8.1
pytz==2019.1
redis==2.10.6
six==1.12.0
sqlparse==0.3.0
traitlets==4.3.2
Twisted==19.7.0
txaio==18.8.1
urllib3==1.25.7
vine==1.3.0
virtualenv==16.7.9
wcwidth==0.1.7
zipp==0.6.0
zope.interface==4.6.0
